+++ front/app/api/graph_data_units/responseTypes/DemographicsWidget.ts
@@ -1,0 +1,19 @@
+export type DemographicsResponse = {
+  data: {
+    type: 'report_builder_data_units';
+    attributes: {
+      series: {
+        [key: string]: number;
+      };
+      // options is undefined for users by birthyear
+      options?: {
+        [key: string]: {
+          title_multiloc: {
+            [key: string]: string;
+          };
+          ordering: number;
+        };
+      };
+    };
+  };
+};
