+++ front/app/api/graph_data_units/responseTypes/index.ts
@@ -52,26 +52,6 @@ export type SingleIdeaResponse = {
   };
 };
 
-export type DemographicsResponse = {
-  data: {
-    type: 'report_builder_data_units';
-    attributes: {
-      series: {
-        [key: string]: number;
-      };
-      // options is undefined for users by birthyear
-      options?: {
-        [key: string]: {
-          title_multiloc: {
-            [key: string]: string;
-          };
-          ordering: number;
-        };
-      };
-    };
-  };
-};
-
 export type GenderOption = 'male' | 'female' | 'unspecified' | '_blank';
 
 export type UsersByGenderResponse = {
